fix: changed numbers type to acocunt for decimals

+++ database/migrations/2024_11_06_150254_create_addresses_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('addresses', function (Blueprint $table) {
             $table->id();
-            $table->smallInteger('longitude');
-            $table->smallInteger('latitude');
+            $table->decimal('longitude', 10, 7);
+            $table->decimal('latitude', 10, 7);
             $table->string('place');
             $table->string('city');
             $table->string('country');
